backend/requirements.txt
aiofiles==24.1.0
aiosqlite==0.17.0
anyio==3.5.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
cmdstanpy==1.2.4
colorama==0.4.6
comm==0.1.2
contourpy==1.3.0
convertdate==2.4.0
cssselect==1.2.0
cycler==0.12.1
Cython==3.0.11
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==4.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
ephem==4.1.5
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fake-useragent==1.5.1
fastapi==0.112.3
fastjsonschema==2.16.2
filelock==3.15.4
Flask==3.0.0
fonttools==4.53.1
fqdn==1.5.1
fsspec==2023.6.0
GeoLiberator==0.3.7
gomaps==0.3.6
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
holidays==0.56
httpcore==1.0.5
httpx==0.27.2
idna==3.4
importlib_metadata==8.0.0
importlib_resources==6.4.4
iniconfig==2.0.0
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.7
LunarCalendar==0.0.9
lxml==5.2.2
MarkupSafe==2.1.2
matplotlib==3.9.2
matplotlib-inline==0.1.6
mistune==2.0.5
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parse==1.20.2
parso==0.8.3
patsy==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pip-system-certs==4.0
pipenv==2024.0.1
platformdirs==4.2.2
playwright==1.45.0
plotly==5.15.0
pluggy==1.3.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
prophet
psutil==5.9.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.8.2
pyee==11.1.0
Pygments==2.14.0
PyMeeus==0.5.12
pyparsing==3.1.4
pyppdf==0.1.2
pyppeteer==2.0.0
pyquery==2.0.0
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.4
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.31.0
requests-html==0.10.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.3.2
scipy==1.14.1
seaborn==0.12.2
selenium==4.10.0
Send2Trash==1.8.0
setuptools-git==1.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.30
sqlparse==0.4.4
stack-data==0.6.2
stanio==0.5.1
starlette==0.38.4
statsmodels==0.13.5
StrEnum==0.4.7
tenacity==9.0.0
terminado==0.17.1
threadpoolctl==3.5.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.8.0
tzdata==2023.3
undetected-chromedriver==3.5.5
uri-template==1.2.0
urllib3==1.26.19
uvicorn==0.30.6
virtualenv==20.26.3
w3lib==2.2.1
wcwidth==0.2.6
webcolors==1.12
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==3.0.1
whitenoise==6.6.0
widgetsnbextension==4.0.5
wrapt==1.11.2
wsproto==1.2.0
zipp==3.19.2
